SALEOR-3825 Update order payment status display resolver
